requirements.txt
# Email Management Agent - Dependencies
# API & Auth
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
msal>=1.24.0
requests>=2.31.0

# Data & NLP
python-dateutil>=2.8.2
regex>=2023.6.3

# Optional: local LLM/summary fallback (no API key)
# anthropic>=0.7.0
# openai>=1.3.0

# Storage & metrics
python-dotenv>=1.0.0

# Web dashboard
flask>=3.0.0
# Production WSGI server (Windows + Linux)
waitress>=3.0.0
